Time: 411 ms (5.11%) | Memory: 17.8 MB (48.18%) - LeetSync

+++ 152-maximum-product-subarray/maximum-product-subarray.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int maxProduct(vector<int>& nums) {
-        int maxi=LONG_MIN;
+        int maxi=INT_MIN;
         int n = nums.size();
         if(n==1) return nums[0];
         for(int i=0; i<n; i++){
